Switch to puma instead of thin, adds nginx configuration.
puma promises an equal or better performance than thin, but
does this with a single master process, which is easier to handle.

The new configuration also serves all http via the backend from
a UNIX socket. It also serves /monit/ via the monit http module.

The nginx configuration is created via "rake nginx:config". Depending
on the Procfile configuration, "rake monit:install" also installs
an entry for the nginx server.