Remove redirection caching.
Looking up a tunnel requires a single database request. However,
a typical usage scenarion would involve a single request to a fqdn, 
after which the user will be redirected away anyway. So there should 
not be many requests to begin with, and then they should appear in long
intervals - so caching seems not to improve things at this time.